move to tracer + suppressed exceptions
